move _process_response in transformation
